moving useTrackContextMenu from Row to BasicKeyframedTrack to fix TS issues
