+++ theatre/studio/src/panels/SequenceEditorPanel/DopeSheet/Right/BasicKeyframedTrack/BasicKeyframedTrack.tsx
@@ -1,19 +1,44 @@
+import React, {useEffect} from 'react'
 import type {TrackData} from '@theatre/core/projects/store/types/SheetState_Historic'
 import type {SequenceEditorPanelLayout} from '@theatre/studio/panels/SequenceEditorPanel/layout/layout'
 import type {SequenceEditorTree_PrimitiveProp} from '@theatre/studio/panels/SequenceEditorPanel/layout/tree'
 import {usePrism} from '@theatre/react'
 import type {Pointer} from '@theatre/dataverse'
 import {val} from '@theatre/dataverse'
-import React from 'react'
 import KeyframeEditor from './KeyframeEditor/KeyframeEditor'
+import useRefAndState from '@theatre/studio/utils/useRefAndState'
+import {useTracksProvider} from '@theatre/studio/panels/SequenceEditorPanel/TracksProvider'
+import {getPasteKeyframesItem} from '@theatre/studio/uiComponents/simpleContextMenu/getCopyPasteKeyframesItem'
+import useContextMenu from '@theatre/studio/uiComponents/simpleContextMenu/useContextMenu'
+import styled from 'styled-components'
+
+const TrackContainer = styled.div<{highlight: boolean}>`
+  height: 100%;
+  position: relative;
+  background: ${(props) => (props.highlight ? '#7a22221f' : 'transparent')};
+`
 
 const BasicKeyframedTrack: React.FC<{
   leaf: SequenceEditorTree_PrimitiveProp
-
   layoutP: Pointer<SequenceEditorPanelLayout>
   trackData: TrackData
 }> = React.memo(({layoutP, trackData, leaf}) => {
-  // TODO: Prevent all the rerenders this causes when selecting keyframes
+  const {trackId} = leaf
+  const {trackToHighlight, setTrackToHighlight} = useTracksProvider()
+  const [ref, refNode] = useRefAndState<HTMLDivElement | null>(null)
+  const [contextMenu, , isOpen] = useTrackContextMenu(refNode, {
+    leaf,
+  })
+
+  useEffect(() => {
+    if (trackId && isOpen) {
+      setTrackToHighlight(trackId)
+    } else {
+      setTrackToHighlight(undefined)
+    }
+  }, [trackId, isOpen])
+
+  // TODO: Prevent the rerenders this causes when selecting keyframes
   const {selectedKeyframeIds, selection} = usePrism(() => {
     const selectionAtom = val(layoutP.selectionAtom)
     const selectedKeyframeIds = val(
@@ -43,7 +68,35 @@ const BasicKeyframedTrack: React.FC<{
     />
   ))
 
-  return <>{keyframeEditors}</>
+  return (
+    <TrackContainer
+      ref={ref}
+      highlight={Boolean(trackId && trackToHighlight === trackId)}
+    >
+      {contextMenu}
+      {keyframeEditors}
+    </TrackContainer>
+  )
 })
+
+function useTrackContextMenu(
+  node: HTMLDivElement | null,
+  {
+    leaf,
+  }: {
+    leaf: SequenceEditorTree_PrimitiveProp
+  },
+) {
+  const pasteKeyframesItem = getPasteKeyframesItem(leaf)
+
+  return useContextMenu(node, {
+    items: () => {
+      if (pasteKeyframesItem) {
+        return [pasteKeyframesItem]
+      }
+      return []
+    },
+  })
+}
 
 export default BasicKeyframedTrack
