+++ theatre/studio/src/panels/SequenceEditorPanel/DopeSheet/Right/Row.tsx
@@ -1,10 +1,6 @@
-import type {SequenceEditorTree_PrimitiveProp} from '@theatre/studio/panels/SequenceEditorPanel/layout/tree'
-import React, {useEffect} from 'react'
+import type {SequenceEditorTree_Row} from '@theatre/studio/panels/SequenceEditorPanel/layout/tree'
+import React from 'react'
 import styled from 'styled-components'
-import useRefAndState from '@theatre/studio/utils/useRefAndState'
-import useContextMenu from '@theatre/studio/uiComponents/simpleContextMenu/useContextMenu'
-import {getPasteKeyframesItem} from '@theatre/studio/uiComponents/simpleContextMenu/getCopyPasteKeyframesItem'
-import {useTracksProvider} from '@theatre/studio/panels/SequenceEditorPanel/TracksProvider'
 
 const Container = styled.li<{}>`
   margin: 0;
@@ -14,7 +10,7 @@ const Container = styled.li<{}>`
   position: relative;
 `
 
-const NodeWrapper = styled.div<{isEven: boolean; highlight: boolean}>`
+const NodeWrapper = styled.div<{isEven: boolean}>`
   box-sizing: border-box;
   width: 100%;
   position: relative;
@@ -30,17 +26,10 @@ const NodeWrapper = styled.div<{isEven: boolean; highlight: boolean}>`
     box-sizing: border-box;
     border-bottom: 1px solid #252b3869;
     background: ${(props) => (props.isEven ? 'transparent' : '#6b8fb505')};
-    transition: background 0.15s ease-in-out;
   }
 
   &:before {
-    transition: background 0.05s ease-in-out;
-    background: ${(props) =>
-      props.highlight
-        ? '#7a22221f'
-        : props.isEven
-        ? 'transparent'
-        : '#6b8fb505'};
+    background: ${(props) => (props.isEven ? 'transparent' : '#6b8fb505')};
   }
 `
 
@@ -51,25 +40,9 @@ const Children = styled.ul`
 `
 
 const Row: React.FC<{
-  leaf: SequenceEditorTree_PrimitiveProp
+  leaf: SequenceEditorTree_Row<unknown>
   node: React.ReactElement
 }> = ({leaf, children, node}) => {
-  const {trackId} = leaf
-  const [ref, refNode] = useRefAndState<HTMLDivElement | null>(null)
-  const [contextMenu, , isOpen] = useTrackContextMenu(refNode, {
-    leaf,
-  })
-
-  const {trackToHighlight, setTrackToHighlight} = useTracksProvider()
-
-  useEffect(() => {
-    if (trackId && isOpen) {
-      setTrackToHighlight(trackId)
-    } else {
-      setTrackToHighlight(undefined)
-    }
-  }, [trackId, isOpen])
-
   const hasChildren = Array.isArray(children) && children.length > 0
 
   return (
@@ -77,35 +50,12 @@ const Row: React.FC<{
       <NodeWrapper
         style={{height: leaf.nodeHeight + 'px'}}
         isEven={leaf.n % 2 === 0}
-        ref={ref}
-        highlight={Boolean(trackId && trackToHighlight === trackId)}
       >
         {node}
       </NodeWrapper>
       {hasChildren && <Children>{children}</Children>}
-      {trackId ? contextMenu : null}
     </Container>
   )
-}
-
-function useTrackContextMenu(
-  node: HTMLDivElement | null,
-  {
-    leaf,
-  }: {
-    leaf: SequenceEditorTree_PrimitiveProp
-  },
-) {
-  const pasteKeyframesItem = getPasteKeyframesItem(leaf)
-
-  return useContextMenu(node, {
-    items: () => {
-      if (pasteKeyframesItem) {
-        return [pasteKeyframesItem]
-      }
-      return []
-    },
-  })
 }
 
 export default Row
